requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
greenlet==1.1.3
idna==3.3
python-decouple==3.6
requests==2.28.1
SQLAlchemy==1.4.40
urllib3==1.26.12
vk-api==11.9.9
